+++ Src/FluentAssertions/Types/TypeSelector.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 using FluentAssertions.Common;
 
 namespace FluentAssertions.Types
@@ -162,6 +163,107 @@ namespace FluentAssertions.Types
         public TypeSelector ThatAreNotUnderNamespace(string @namespace)
         {
             types = types.Where(t => !t.IsUnderNamespace(@namespace)).ToList();
+            return this;
+        }
+
+        /// <summary>
+        /// Determines whether the type is a class
+        /// </summary>
+        public TypeSelector ThatAreClasses()
+        {
+            types = types.Where(t => t.IsClass).ToList();
+            return this;
+        }
+
+        /// <summary>
+        /// Determines whether the type is not a class
+        /// </summary>
+        public TypeSelector ThatAreNotClasses()
+        {
+            types = types.Where(t => !t.IsClass).ToList();
+            return this;
+        }
+
+        /// <summary>
+        /// Determines whether the type is static
+        /// </summary>
+        public TypeSelector ThatAreStatic()
+        {
+            types = types.Where(t => t.IsCSharpStatic()).ToList();
+            return this;
+        }
+
+        /// <summary>
+        /// Determines whether the type is not static
+        /// </summary>
+        public TypeSelector ThatAreNotStatic()
+        {
+            types = types.Where(t => !t.IsCSharpStatic()).ToList();
+            return this;
+        }
+
+        /// <summary>
+        /// Allows to filter the types with the <paramref name="predicate"/> passed
+        /// </summary>
+        public TypeSelector ThatSatisfy(Func<Type, bool> predicate)
+        {
+            types = types.Where(predicate).ToList();
+            return this;
+        }
+
+        /// <summary>
+        /// Returns T for the types which are <see cref="Task{T}"/> or <see cref="ValueTask{T}"/>; the type itself otherwise
+        /// </summary>
+        public TypeSelector UnwrapTaskTypes()
+        {
+            types = types.Select(type =>
+            {
+                if (type.IsGenericType && type.GetGenericTypeDefinition() == typeof(Task<>))
+                {
+                    return type.GetGenericArguments().Single();
+                }
+                if (type.IsGenericType && type.GetGenericTypeDefinition() == typeof(ValueTask<>))
+                {
+                    return type.GetGenericArguments().Single();
+                }
+                return type == typeof(Task) || type == typeof(ValueTask) ? typeof(void) : type;
+            }).ToList();
+
+            return this;
+        }
+
+        /// <summary>
+        /// Returns T for the types which are <see cref="IEnumerable{T}"/> or implement the <see cref="IEnumerable{T}"/>; the type itself otherwise
+        /// </summary>
+        public TypeSelector UnwrapEnumerableTypes()
+        {
+            var unwrappedTypes = new List<Type>();
+            foreach (Type type in types)
+            {
+                if (type.IsGenericType && type.GetGenericTypeDefinition() == typeof(IEnumerable<>))
+                {
+                    unwrappedTypes.Add(type.GetGenericArguments().Single());
+                }
+                else
+                {
+                    var iEnumerableImplementations = type
+                        .GetInterfaces()
+                        .Where(iType => iType.IsGenericType && iType.GetGenericTypeDefinition() == typeof(IEnumerable<>))
+                        .Select(ied => ied.GetGenericArguments().Single())
+                        .ToList();
+
+                    if (iEnumerableImplementations.Any())
+                    {
+                        unwrappedTypes.AddRange(iEnumerableImplementations);
+                    }
+                    else
+                    {
+                        unwrappedTypes.Add(type);
+                    }
+                }
+            }
+
+            types = unwrappedTypes;
             return this;
         }