+++ Tests/FluentAssertions.Specs/Types/TypeSelectorSpecs.cs
@@ -1,13 +1,18 @@
 ﻿using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Reflection;
-
+using System.Threading.Tasks;
 using FluentAssertions.Types;
 
 using Internal.Main.Test;
+using Internal.NotOnlyClasses.Test;
 using Internal.Other.Test;
 using Internal.Other.Test.Common;
+using Internal.StaticAndNonStaticClasses.Test;
+using Internal.UnwrapSelectorTestTypes.Test;
 using Xunit;
+using ISomeInterface = Internal.Main.Test.ISomeInterface;
 
 namespace FluentAssertions.Specs
 {
@@ -415,6 +420,119 @@ namespace FluentAssertions.Specs
             // Assert
             filteredTypes.As<IEnumerable<Type>>().Should().ContainSingle();
         }
+
+        [Fact]
+        public void When_selecting_types_that_are_classes_it_should_return_the_correct_types()
+        {
+            // Arrange
+            TypeSelector types = new[] { typeof(NotOnlyClassesClass), typeof(NotOnlyClassesEnumeration), typeof(INotOnlyClassesInterface) }.Types();
+
+            // Act
+            IEnumerable<Type> filteredTypes = types.ThatAreClasses();
+
+            // Assert
+            filteredTypes.Should()
+                .ContainSingle()
+                .Which.Should().Be(typeof(NotOnlyClassesClass));
+        }
+
+        [Fact]
+        public void When_selecting_types_that_are_not_classes_it_should_return_the_correct_types()
+        {
+            // Arrange
+            Assembly assembly = typeof(NotOnlyClassesClass).GetTypeInfo().Assembly;
+
+            // Act
+            IEnumerable<Type> types = AllTypes.From(assembly)
+                .ThatAreInNamespace("Internal.NotOnlyClasses.Test")
+                .ThatAreNotClasses();
+
+            // Assert
+            types.Should()
+                .HaveCount(2)
+                .And.Contain(typeof(INotOnlyClassesInterface))
+                .And.Contain(typeof(NotOnlyClassesEnumeration));
+        }
+
+        [Fact]
+        public void When_selecting_types_that_are_static_classes_it_should_return_the_correct_types()
+        {
+            // Arrange
+            Assembly assembly = typeof(StaticClass).GetTypeInfo().Assembly;
+
+            // Act
+            IEnumerable<Type> types = AllTypes.From(assembly)
+                .ThatAreInNamespace("Internal.StaticAndNonStaticClasses.Test")
+                .ThatAreStatic();
+
+            // Assert
+            types.Should()
+                .ContainSingle()
+                .Which.Should().Be(typeof(StaticClass));
+        }
+
+        [Fact]
+        public void When_selecting_types_that_are_not_static_classes_it_should_return_the_correct_types()
+        {
+            // Arrange
+            Assembly assembly = typeof(StaticClass).GetTypeInfo().Assembly;
+
+            // Act
+            IEnumerable<Type> types = AllTypes.From(assembly)
+                .ThatAreInNamespace("Internal.StaticAndNonStaticClasses.Test")
+                .ThatAreNotStatic();
+
+            // Assert
+            types.Should()
+                .ContainSingle()
+                .Which.Should().Be(typeof(NotAStaticClass));
+        }
+
+        [Fact]
+        public void When_selecting_types_with_predicate_it_should_return_the_correct_types()
+        {
+            // Arrange
+            Assembly assembly = typeof(SomeBaseClass).GetTypeInfo().Assembly;
+
+            // Act
+            IEnumerable<Type> types = AllTypes.From(assembly)
+                .ThatSatisfy(t => t.GetCustomAttribute<SomeAttribute>() != null);
+
+            // Assert
+            types.Should()
+                .HaveCount(3)
+                .And.Contain(typeof(ClassWithSomeAttribute))
+                .And.Contain(typeof(ClassWithSomeAttributeDerived))
+                .And.Contain(typeof(ClassWithSomeAttributeThatImplementsSomeInterface));
+        }
+        
+        [Fact]
+        public void When_unwrap_task_types_it_should_return_the_correct_types()
+        {
+            IEnumerable<Type> types = typeof(ClassToExploreUnwrappedTaskTypes)
+                .Methods()
+                .ReturnTypes()
+                .UnwrapTaskTypes();
+
+            types.Should()
+                .BeEquivalentTo(typeof(int), typeof(void), typeof(void), typeof(string), typeof(bool));
+        }
+
+        [Fact]
+        public void When_unwrap_enumerable_types_it_should_return_the_correct_types()
+        {
+            IEnumerable<Type> types = typeof(ClassToExploreUnwrappedEnumerableTypes)
+                .Methods()
+                .ReturnTypes()
+                .UnwrapEnumerableTypes();
+
+            types.Should()
+                .HaveCount(4)
+                .And.Contain(typeof(IEnumerable))
+                .And.Contain(typeof(bool))
+                .And.Contain(typeof(int))
+                .And.Contain(typeof(string));
+        }
     }
 }
 
@@ -507,6 +625,63 @@ namespace Internal.Other.Test.Common
 {
     internal class SomeCommonClass
     {
+    }
+}
+
+namespace Internal.NotOnlyClasses.Test
+{
+    internal class NotOnlyClassesClass
+    {
+    }
+
+    internal enum NotOnlyClassesEnumeration
+    {
+    }
+
+    internal interface INotOnlyClassesInterface
+    {
+    }
+}
+
+namespace Internal.StaticAndNonStaticClasses.Test
+{
+    internal static class StaticClass
+    {
+    }
+
+    internal class NotAStaticClass
+    {
+    }
+}
+
+namespace Internal.UnwrapSelectorTestTypes.Test
+{
+    internal class ClassToExploreUnwrappedTaskTypes
+    {
+        internal int DoWithInt() { return default; }
+        internal Task DoWithTask() { return Task.CompletedTask; }
+        internal ValueTask DoWithValueTask() { return new ValueTask(); }
+        internal Task<string> DoWithIntTask() { return Task.FromResult(string.Empty); }
+        internal ValueTask<bool> DoWithBoolValueTask() { return new ValueTask<bool>(false); }
+    }
+
+    internal class ClassToExploreUnwrappedEnumerableTypes
+    {
+        internal IEnumerable DoWithTask() { return default; }
+        internal List<bool> DoWithIntTask() { return default; }
+        internal ClassImplementingMultipleEnumerable DoWithBoolValueTask() { return default; }
+    }
+
+    internal class ClassImplementingMultipleEnumerable : IEnumerable<int>, IEnumerable<string>
+    {
+        private readonly IEnumerable<int> integers = new int[0];
+        private readonly IEnumerable<string> strings = new string[0];
+
+        public IEnumerator<int> GetEnumerator() => integers.GetEnumerator();
+
+        IEnumerator IEnumerable.GetEnumerator() => ((IEnumerable)integers).GetEnumerator();
+
+        IEnumerator<string> IEnumerable<string>.GetEnumerator() => strings.GetEnumerator();
     }
 }